pyproject.toml
[project]
version = "1.0.0"
name = "{{ project_name }}_server"
package-mode = false
requires-python = "==3.13.3"
dependencies = [
    "dj-database-url==2.3.0",
    "django==5.2.1",
    "django-cors-headers==4.7.0",
    "gunicorn==23.0.0",
    "psycopg[binary]==3.2.9",
    "requests==2.32.3",
    "python-dotenv==1.1.0",
    "djangorestframework==3.16.0",
    "drf_spectacular==0.28.0",
    "django-cache-url==3.4.5",
    "sentry-sdk==2.28.0",
]
